Remove "press q to quit" altogether
The logic in previous commit was utterly silly. Of course the pager is
always still running if we can write a helping message to it! :)

On hindsight remove the quit-help altogether. I think we can trust
people who use a tool such as ghi to know their pager. Besides it's
quite easy to get less to quit.